Introduces RSpec and Travis
This commit adds in Rspec, which will be used as the main testing suite.
As such the default test directory was removed as rspec instead places
its important files inside of the spec directory. No tests were created
as the app is such an infant and there isn't really anything to test.

This commit also adds in the travis and travis lint gems and introduces
a rather basic .travis.yml file. The travis.yml file was created from
the command line by running travis init.

This created a pretty paired down .travis.yml file that was essentially
no good. Follow up and play testing found that adding sudo and ruby v 2.3.1
was a good idea had builds passing.

I also added a slew of gems I found online that seemed helpful and wrote a
baby script that runs the full test suite. In the future we will want to
make this a bit more sophisticated to run different parts of the test suite
in different jobs so we can quickly see "oh I broke a unit test" or what have
you.

Travis was added as it is an awesome tool that will soon auto-deploy
builds from master into the Heroku project after they have passed
the test suite.