Adds board to game#show page
This game board is built off of css and flexbox.

The board displays the rows in a column (vertical) direction.
It has a set max and min width and responds with width 90% on
window resize.

The rows display the squares in a row (left-to-right) direction.

Squares are styled based upon their nth:child(odd) or even status
as appropriate to create a checker board effect. Flex 1 has them
filling the whole space granted to them including on window rezise
to create a responsive effect.

As equating height to width is difficult to do through css alone
I added an equalHeight jQuery function which resizes the squares
to match the height both initially and upon window resize.