Rebuilds app to have players instead of users
This is a giant of a commit and would have ideally been broken down into
more bite size pieces.

That being said there is a strong theme here. Where there were users,
there is now only players. This was done to have the app's syntax flow
better.

In addition, there was some serious database schema changes. As the app
is so new, I went ahead a just rewrote the initial migrations rather
than having a slew of more confusing drop this and this and this
migrations.

As such you will need to run
  - rake db:drop
  - rake db:create
  - rake db:migrate

to have your local db be in sync with what's now standard.