Rewrites piece moving methods
Attempt move steps through several situations to see if the piece
can and should move to the new location. If it can it then makes
the move and increments the game's turn count.

This makes the capture logic simple and updates syntax to ensure that
first move is set to false after a move has happened.